Refactor imagebufalgo.cpp -- too much code for one file, split it into parts * imagebufalgo_compare.cpp - analysis and comparison of images * imagebufalgo_copy.cpp - move pixels or channels around, don't change values * imagebufalgo_pixelmath.cpp -- math on single pixels at a time Everything not falling into these categories was left in imagebufalgo.cpp. This nicely splits the IBA functions almost evenly between the files, which will make much kinder compile times when tinkering with any one.
