+++ src/libOpenImageIO/CMakeLists.txt
@@ -35,6 +35,9 @@ set (libOpenImageIO_srcs exif.cpp formatspec.cpp imagebuf.cpp
                           imageoutput.cpp iptc.cpp xmp.cpp
                           color_ocio.cpp
                           imagebufalgo.cpp
+                          imagebufalgo_compare.cpp
+                          imagebufalgo_copy.cpp
+                          imagebufalgo_pixelmath.cpp
                           imagebufalgo_yee.cpp imagebufalgo_opencv.cpp
                           maketexture.cpp
                           ../libutil/argparse.cpp
