+++ src/libOpenImageIO/imagebufalgo.cpp
@@ -28,20 +28,12 @@
   (This is the Modified BSD License)
 */
 
-/* This header has to be included before boost/regex.hpp header
-   If it is included after, there is an error
-   "undefined reference to CSHA1::Update (unsigned char const*, unsigned long)"
-*/
-#include "SHA1.h"
-
 /// \file
 /// Implementation of ImageBufAlgo algorithms.
 
-#include <boost/version.hpp>
 #include <boost/bind.hpp>
 #include <boost/shared_ptr.hpp>
 
-#include <OpenEXR/ImathFun.h>
 #include <OpenEXR/half.h>
 
 #include <cmath>
@@ -63,9 +55,6 @@
 #include FT_FREETYPE_H
 #endif
 
-#ifdef USE_OPENSSL
-#include <openssl/sha.h>
-#endif
 
 ///////////////////////////////////////////////////////////////////////////
 // Guidelines for ImageBufAlgo functions:
@@ -269,1535 +258,6 @@ ImageBufAlgo::checker (ImageBuf &dst, int width,
     return checker (dst, width, width, width, color1, color2, 0, 0, 0,
                     ROI(xbegin,xend,ybegin,yend,zbegin,zend), 0);
 }
-
-
-
-template<class D, class S>
-static bool
-paste_ (ImageBuf &dst, ROI dstroi,
-        const ImageBuf &src, ROI srcroi, int nthreads)
-{
-    // N.B. Punt on parallelizing because of the subtle interplay
-    // between srcroi and dstroi, the parallel_image idiom doesn't
-    // handle that especially well. And it's not worth customizing for
-    // this function which is inexpensive and not commonly used, and so
-    // would benefit little from parallelizing. We can always revisit
-    // this later. But in the mean time, we maintain the 'nthreads'
-    // parameter for uniformity with the rest of IBA.
-    int src_nchans = src.nchannels ();
-    int dst_nchans = dst.nchannels ();
-    ImageBuf::ConstIterator<S,D> s (src, srcroi);
-    ImageBuf::Iterator<D,D> d (dst, dstroi);
-    for ( ;  ! s.done();  ++s, ++d) {
-        if (! d.exists())
-            continue;  // Skip paste-into pixels that don't overlap dst's data
-        for (int c = srcroi.chbegin, c_dst = dstroi.chbegin;
-             c < srcroi.chend;  ++c, ++c_dst) {
-            if (c_dst >= 0 && c_dst < dst_nchans)
-                d[c_dst] = c < src_nchans ? s[c] : D(0);
-        }
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::paste (ImageBuf &dst, int xbegin, int ybegin,
-                     int zbegin, int chbegin,
-                     const ImageBuf &src, ROI srcroi, int nthreads)
-{
-    if (! srcroi.defined())
-        srcroi = get_roi(src.spec());
-
-    ROI dstroi (xbegin, xbegin+srcroi.width(),
-                ybegin, ybegin+srcroi.height(),
-                zbegin, zbegin+srcroi.depth(),
-                chbegin, chbegin+srcroi.nchannels());
-    ROI dstroi_save = dstroi;  // save the original
-    IBAprep (dstroi, &dst);
-
-    // do the actual copying
-    OIIO_DISPATCH_TYPES2 ("paste", paste_, dst.spec().format, src.spec().format,
-                          dst, dstroi_save, src, srcroi, nthreads);
-    return false;
-}
-
-
-
-
-template<class D, class S>
-static bool
-crop_ (ImageBuf &dst, const ImageBuf &src,
-       ROI roi, int nthreads=1)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Lots of pixels and request for multi threads? Parallelize.
-        ImageBufAlgo::parallel_image (
-            boost::bind(crop_<D,S>, boost::ref(dst), boost::cref(src),
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-    ImageBuf::ConstIterator<S,D> s (src, roi);
-    ImageBuf::Iterator<D,D> d (dst, roi);
-    for ( ;  ! d.done();  ++d, ++s) {
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            d[c] = s[c];
-    }
-    return true;
-}
-
-
-
-bool 
-ImageBufAlgo::crop (ImageBuf &dst, const ImageBuf &src,
-                    ROI roi, int nthreads)
-{
-    dst.clear ();
-    roi.chend = std::min (roi.chend, src.nchannels());
-    IBAprep (roi, &dst, &src);
-    OIIO_DISPATCH_TYPES2 ("crop", crop_, dst.spec().format, src.spec().format,
-                          dst, src, roi, nthreads);
-    return false;
-}
-
-
-
-template<class D, class S>
-static bool
-flip_ (ImageBuf &dst, const ImageBuf &src, ROI roi, int nthreads)
-{
-    ImageBuf::ConstIterator<S, D> s (src, roi);
-    ImageBuf::Iterator<D, D> d (dst, roi);
-    for ( ; ! d.done(); ++d) {
-        s.pos (d.x(), roi.yend-1 - (d.y() - roi.ybegin), d.z());
-        for (int c = roi.chbegin; c < roi.chend; ++c)
-            d[c] = s[c];
-    }
-    return true;
-}
-
-
-bool
-ImageBufAlgo::flip(ImageBuf &dst, const ImageBuf &src, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES2 ("flip", flip_,
-                          dst.spec().format, src.spec().format, dst, src,
-                          roi, nthreads);
-    return false;
-}
-
-
-
-template<class D, class S>
-static bool
-flop_ (ImageBuf &dst, const ImageBuf &src, ROI roi, int nthreads)
-{
-    ImageBuf::ConstIterator<S, D> s (src, roi);
-    ImageBuf::Iterator<D, D> d (dst, roi);
-    for ( ; ! d.done(); ++d) {
-        s.pos (roi.xend-1 - (d.x() - roi.xbegin), d.y(), d.z());
-        for (int c = roi.chbegin; c < roi.chend; ++c)
-            d[c] = s[c];
-    }
-    return true;
-}
-
-
-bool
-ImageBufAlgo::flop (ImageBuf &dst, const ImageBuf &src, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES2 ("flop", flop_,
-                          dst.spec().format, src.spec().format, dst, src,
-                          roi, nthreads);
-    return false;
-}
-
-
-
-template<class D, class S>
-static bool
-flipflop_ (ImageBuf &dst, const ImageBuf &src, ROI roi, int nthreads)
-{
-    // Serial case
-    ImageBuf::ConstIterator<S, D> s (src, roi);
-    ImageBuf::Iterator<D, D> d (dst, roi);
-    for ( ; !d.done(); ++d) {
-        s.pos (roi.xend-1 - (d.x() - roi.xbegin),
-               roi.yend-1 - (d.y() - roi.ybegin), d.z());
-        for (int c = roi.chbegin; c < roi.chend; ++c)
-            d[c] = s[c];
-    }
-    return true;
-}
-
-
-bool
-ImageBufAlgo::flipflop (ImageBuf &dst, const ImageBuf &src,
-                        ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES2 ("flipflop", flipflop_,
-                          dst.spec().format, src.spec().format, dst, src,
-                          roi, nthreads);
-    return false;
-}
-
-
-
-template<class D>
-static bool
-clamp_ (ImageBuf &dst, const float *min, const float *max,
-        bool clampalpha01, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Lots of pixels and request for multi threads? Parallelize.
-        ImageBufAlgo::parallel_image (
-            boost::bind(clamp_<D>, boost::ref(dst), min, max, clampalpha01,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-    for (ImageBuf::Iterator<D> d (dst, roi);  ! d.done();  ++d) {
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            d[c] = OIIO::clamp<float> (d[c], min[c], max[c]);
-    }
-    int a = dst.spec().alpha_channel;
-    if (clampalpha01 && a >= roi.chbegin && a < roi.chend) {
-        for (ImageBuf::Iterator<D> d (dst, roi);  ! d.done();  ++d) {
-            d[a] = OIIO::clamp<float> (d[a], 0.0f, 1.0f);
-        }
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::clamp (ImageBuf &dst, const float *min, const float *max,
-                     bool clampalpha01, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    std::vector<float> minvec, maxvec;
-    if (! min) {
-        minvec.resize (dst.nchannels(), -std::numeric_limits<float>::max());
-        min = &minvec[0];
-    }
-    if (! max) {
-        maxvec.resize (dst.nchannels(), std::numeric_limits<float>::max());
-        max = &maxvec[0];
-    }
-    OIIO_DISPATCH_TYPES ("clamp", clamp_, dst.spec().format, dst,
-                         min, max, clampalpha01, roi, nthreads);
-    return false;
-}
-
-
-
-bool
-ImageBufAlgo::clamp (ImageBuf &dst, float min, float max,
-                     bool clampalpha01, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    std::vector<float> minvec (dst.nchannels(), min);
-    std::vector<float> maxvec (dst.nchannels(), max);
-    OIIO_DISPATCH_TYPES ("clamp", clamp_, dst.spec().format, dst,
-                         &minvec[0], &maxvec[0], clampalpha01, roi, nthreads);
-    return false;
-}
-
-
-
-bool
-ImageBufAlgo::channels (ImageBuf &dst, const ImageBuf &src,
-                        int nchannels, const int *channelorder,
-                        bool shuffle_channel_names)
-{
-    // DEPRECATED -- just provide link compatibility
-    return channels (dst, src, nchannels, channelorder, NULL, NULL,
-                     shuffle_channel_names);
-}
-
-
-
-bool
-ImageBufAlgo::channels (ImageBuf &dst, const ImageBuf &src,
-                        int nchannels, const int *channelorder,
-                        const float *channelvalues,
-                        const std::string *newchannelnames,
-                        bool shuffle_channel_names)
-{
-    // Not intended to create 0-channel images.
-    if (nchannels <= 0) {
-        dst.error ("%d-channel images not supported", nchannels);
-        return false;
-    }
-    // If we dont have a single source channel,
-    // hard to know how big to make the additional channels
-    if (src.spec().nchannels == 0) {
-        dst.error ("%d-channel images not supported", src.spec().nchannels);
-        return false;
-    }
-
-    // If channelorder is NULL, it will be interpreted as
-    // {0, 1, ..., nchannels-1}.
-    int *local_channelorder = NULL;
-    if (! channelorder) {
-        local_channelorder = ALLOCA (int, nchannels);
-        for (int c = 0;  c < nchannels;  ++c)
-            local_channelorder[c] = c;
-        channelorder = local_channelorder;
-    }
-
-    // If this is the identity transformation, just do a simple copy
-    bool inorder = true;
-    for (int c = 0;  c < nchannels;   ++c)
-        inorder &= (channelorder[c] == c);
-    if (nchannels == src.spec().nchannels && inorder) {
-        return dst.copy (src);
-    }
-
-    // Construct a new ImageSpec that describes the desired channel ordering.
-    ImageSpec newspec = src.spec();
-    newspec.nchannels = nchannels;
-    newspec.default_channel_names ();
-    newspec.alpha_channel = -1;
-    newspec.z_channel = -1;
-    for (int c = 0; c < nchannels;  ++c) {
-        int csrc = channelorder[c];
-        // If the user gave an explicit name for this channel, use it...
-        if (newchannelnames && newchannelnames[c].size())
-            newspec.channelnames[c] = newchannelnames[c];
-        // otherwise, if shuffle_channel_names, use the channel name of
-        // the src channel we're using (otherwise stick to the default name)
-        else if (shuffle_channel_names &&
-                 csrc >= 0 && csrc < src.spec().nchannels)
-            newspec.channelnames[c] = src.spec().channelnames[csrc];
-        // otherwise, use the name of the source in that slot
-        else if (csrc >= 0 && csrc < src.spec().nchannels)
-            newspec.channelnames[c] = src.spec().channelnames[c];
-        // Use the names (or designation of the src image, if
-        // shuffle_channel_names is true) to deduce the alpha and z channels.
-        if ((shuffle_channel_names && csrc == src.spec().alpha_channel) ||
-              Strutil::iequals (newspec.channelnames[c], "A") ||
-              Strutil::iequals (newspec.channelnames[c], "alpha"))
-            newspec.alpha_channel = c;
-        if ((shuffle_channel_names && csrc == src.spec().z_channel) ||
-              Strutil::iequals (newspec.channelnames[c], "Z"))
-            newspec.z_channel = c;
-    }
-
-    // Update the image (realloc with the new spec)
-    dst.alloc (newspec);
-
-    // Copy the channels individually
-    stride_t dstxstride = AutoStride, dstystride = AutoStride, dstzstride = AutoStride;
-    ImageSpec::auto_stride (dstxstride, dstystride, dstzstride,
-                            newspec.format.size(), newspec.nchannels,
-                            newspec.width, newspec.height);
-    int channelsize = newspec.format.size();
-    char *pixels = (char *) dst.pixeladdr (dst.xbegin(), dst.ybegin(),
-                                           dst.zbegin());
-    for (int c = 0;  c < nchannels;  ++c) {
-        // Copy shuffled channels
-        if (channelorder[c] >= 0 && channelorder[c] < src.spec().nchannels) {
-            int csrc = channelorder[c];
-            src.get_pixel_channels (src.xbegin(), src.xend(),
-                                    src.ybegin(), src.yend(),
-                                    src.zbegin(), src.zend(),
-                                    csrc, csrc+1, newspec.format, pixels,
-                                    dstxstride, dstystride, dstzstride);
-        }
-        // Set channels that are literals
-        if (channelorder[c] < 0 && channelvalues && channelvalues[c]) {
-            ROI roi = get_roi (dst.spec());
-            roi.chbegin = c;
-            roi.chend = c+1;
-            ImageBufAlgo::fill (dst, &channelvalues[0], roi);
-        }
-        pixels += channelsize;
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::setNumChannels(ImageBuf &dst, const ImageBuf &src, int numChannels)
-{
-    return ImageBufAlgo::channels (dst, src, numChannels, NULL, NULL, NULL, true);
-}
-
-
-
-template<class ABtype>
-static bool
-channel_append_impl (ImageBuf &dst, const ImageBuf &A, const ImageBuf &B,
-                     ROI roi, int nthreads)
-{
-    if (nthreads == 1 || roi.npixels() < 1000) {
-        int na = A.nchannels(), nb = B.nchannels();
-        int n = std::min (dst.nchannels(), na+nb);
-        ImageBuf::Iterator<float> r (dst, roi);
-        ImageBuf::ConstIterator<ABtype> a (A, roi);
-        ImageBuf::ConstIterator<ABtype> b (B, roi);
-        for (;  !r.done();  ++r) {
-            a.pos (r.x(), r.y(), r.z());
-            b.pos (r.x(), r.y(), r.z());
-            for (int c = 0; c < n; ++c) {
-                if (c < na)
-                    r[c] = a.exists() ? a[c] : 0.0f;
-                else
-                    r[c] = b.exists() ? b[c-na] : 0.0f;
-            }
-        }
-    } else {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind (channel_append_impl<ABtype>, boost::ref(dst),
-                         boost::cref(A), boost::cref(B), _1, 1),
-            roi, nthreads);
-    }
-    return true;
-}
-
-
-bool
-ImageBufAlgo::channel_append (ImageBuf &dst, const ImageBuf &A,
-                              const ImageBuf &B, ROI roi,
-                              int nthreads)
-{
-    // If the region is not defined, set it to the union of the valid
-    // regions of the two source images.
-    if (! roi.defined())
-        roi = roi_union (get_roi (A.spec()), get_roi (B.spec()));
-
-    // If dst has not already been allocated, set it to the right size,
-    // make it unconditinally float.
-    if (! dst.pixels_valid()) {
-        ImageSpec dstspec = A.spec();
-        dstspec.set_format (TypeDesc::TypeFloat);
-        // Append the channel descriptions
-        dstspec.nchannels = A.spec().nchannels + B.spec().nchannels;
-        for (int c = 0;  c < B.spec().nchannels;  ++c) {
-            std::string name = B.spec().channelnames[c];
-            // Eliminate duplicates
-            if (std::find(dstspec.channelnames.begin(), dstspec.channelnames.end(), name) != dstspec.channelnames.end())
-                name = Strutil::format ("channel%d", A.spec().nchannels+c);
-            dstspec.channelnames.push_back (name);
-        }
-        if (dstspec.alpha_channel < 0 && B.spec().alpha_channel >= 0)
-            dstspec.alpha_channel = B.spec().alpha_channel + A.nchannels();
-        if (dstspec.z_channel < 0 && B.spec().z_channel >= 0)
-            dstspec.z_channel = B.spec().z_channel + A.nchannels();
-        set_roi (dstspec, roi);
-        dst.reset (dst.name(), dstspec);
-    }
-
-    // For now, only support float destination, and equivalent A and B
-    // types.
-    if (dst.spec().format != TypeDesc::FLOAT ||
-        A.spec().format != B.spec().format) {
-        dst.error ("Unable to perform channel_append of %s, %s -> %s",
-                   A.spec().format, B.spec().format, dst.spec().format);
-        return false;
-    }
-
-    OIIO_DISPATCH_TYPES ("channel_append", channel_append_impl,
-                         A.spec().format, dst, A, B, roi, nthreads);
-    return true;
-}
-
-
-
-// DEPRECATED version
-bool
-ImageBufAlgo::add (ImageBuf &dst, const ImageBuf &A, const ImageBuf &B,
-                   int options)
-{
-    // Sanity checks
-    
-    // dst must be distinct from A and B
-    if ((const void *)&A == (const void *)&dst ||
-        (const void *)&B == (const void *)&dst) {
-        dst.error ("destination image must be distinct from source");
-        return false;
-    }
-    
-    // all three images must have the same number of channels
-    if (A.spec().nchannels != B.spec().nchannels) {
-        dst.error ("channel number mismatch: %d vs. %d", 
-                   A.spec().nchannels, B.spec().nchannels);
-        return false;
-    }
-    
-    // If dst has not already been allocated, set it to the right size,
-    // make it unconditinally float
-    if (! dst.pixels_valid()) {
-        ImageSpec dstspec = A.spec();
-        dstspec.set_format (TypeDesc::TypeFloat);
-        dst.alloc (dstspec);
-    }
-    // Clear dst pixels if instructed to do so
-    if (options & ADD_CLEAR_DST) {
-        zero (dst);
-    }
-      
-    ASSERT (A.spec().format == TypeDesc::FLOAT &&
-            B.spec().format == TypeDesc::FLOAT &&
-            dst.spec().format == TypeDesc::FLOAT);
-    
-    ImageBuf::ConstIterator<float,float> a (A);
-    ImageBuf::ConstIterator<float,float> b (B);
-    ImageBuf::Iterator<float> d (dst);
-    int nchannels = A.nchannels();
-    // Loop over all pixels in A
-    for ( ; a.valid();  ++a) {  
-        // Point the iterators for B and dst to the corresponding pixel
-        if (options & ADD_RETAIN_WINDOWS) {
-            b.pos (a.x(), a.y());
-        } else {
-            // ADD_ALIGN_WINDOWS: make B line up with A
-            b.pos (a.x()-A.xbegin()+B.xbegin(), a.y()-A.ybegin()+B.ybegin());
-        }
-        d.pos (a.x(), b.y());
-        
-        if (! b.valid() || ! d.valid())
-            continue;   // Skip pixels that don't align
-        
-        // Add the pixel
-        for (int c = 0;  c < nchannels;  ++c)
-              d[c] = a[c] + b[c];
-    }
-    
-    return true;
-}
-
-
-
-template<class Rtype, class Atype, class Btype>
-static bool
-add_impl (ImageBuf &R, const ImageBuf &A, const ImageBuf &B,
-          ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(add_impl<Rtype,Atype,Btype>,
-                        boost::ref(R), boost::cref(A), boost::cref(B),
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    ImageBuf::ConstIterator<Atype> a (A, roi);
-    ImageBuf::ConstIterator<Btype> b (B, roi);
-    for ( ;  !r.done();  ++r, ++a, ++b)
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            r[c] = a[c] + b[c];
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::add (ImageBuf &dst, const ImageBuf &A, const ImageBuf &B,
-                   ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst, &A, &B);
-    OIIO_DISPATCH_COMMON_TYPES3 ("add", add_impl, dst.spec().format,
-                                 A.spec().format, B.spec().format,
-                                 dst, A, B, roi, nthreads);
-    return true;
-}
-
-
-
-template<class Rtype>
-static bool
-add_inplace (ImageBuf &R, const float *val,
-             ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(add_inplace<Rtype>, boost::ref(R), val,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    for ( ;  !r.done();  ++r)
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            r[c] = r[c] + val[c];
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::add (ImageBuf &dst, const float *val, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES ("add", add_inplace, dst.spec().format,
-                         dst, val, roi, nthreads);
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::add (ImageBuf &R, float val, ROI roi, int nthreads)
-{
-    int nc = R.nchannels();
-    float *vals = ALLOCA (float, nc);
-    for (int c = 0;  c < nc;  ++c)
-        vals[c] = val;
-    return add (R, vals, roi, nthreads);
-}
-
-
-
-
-template<class Rtype, class Atype, class Btype>
-static bool
-sub_impl (ImageBuf &R, const ImageBuf &A, const ImageBuf &B,
-          ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(sub_impl<Rtype,Atype,Btype>,
-                        boost::ref(R), boost::cref(A), boost::cref(B),
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    ImageBuf::ConstIterator<Atype> a (A, roi);
-    ImageBuf::ConstIterator<Btype> b (B, roi);
-    for ( ;  !r.done();  ++r, ++a, ++b)
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            r[c] = a[c] - b[c];
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::sub (ImageBuf &dst, const ImageBuf &A, const ImageBuf &B,
-                   ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst, &A, &B);
-    OIIO_DISPATCH_COMMON_TYPES3 ("sub", sub_impl, dst.spec().format,
-                                 A.spec().format, B.spec().format,
-                                 dst, A, B, roi, nthreads);
-    return true;
-}
-
-
-
-template<class Rtype>
-static bool
-mul_impl (ImageBuf &R, const float *val, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(mul_impl<Rtype>, boost::ref(R), val,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    for (ImageBuf::Iterator<Rtype> r (R, roi);  !r.done();  ++r)
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            r[c] = r[c] * val[c];
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::mul (ImageBuf &dst, const float *val, ROI roi, int nthreads)
-{
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES ("mul", mul_impl, dst.spec().format,
-                         dst, val, roi, nthreads);
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::mul (ImageBuf &R, float val, ROI roi, int nthreads)
-{
-    int nc = R.nchannels();
-    float *vals = ALLOCA (float, nc);
-    for (int c = 0;  c < nc;  ++c)
-        vals[c] = val;
-    return mul (R, vals, roi, nthreads);
-}
-
-
-
-inline float rangecompress (float x)
-{
-    // Formula courtesy of Sony Pictures Imageworks
-    const float x1 = 1.0, a = 1.2607481479644775391;
-    const float b = 0.28785100579261779785, c = -1.4042005538940429688;
-    float absx = fabsf(x);
-    if (absx <= x1)
-        return x;
-    return copysignf (a + b * logf(fabsf(c*absx + 1.0f)), x);
-}
-
-
-
-inline float rangeexpand (float y)
-{
-    // Formula courtesy of Sony Pictures Imageworks
-    const float x1 = 1.0, a = 1.2607481479644775391;
-    const float b = 0.28785100579261779785, c = -1.4042005538940429688;
-    float absy = fabsf(y);
-    if (absy <= x1)
-        return y;
-    float xIntermediate = expf ((absy - a)/b);
-    // Since the compression step includes an absolute value, there are
-    // two possible results here. If x < x1 it is the incorrect result,
-    // so pick the other value.
-    float x = (xIntermediate - 1.0f) / c;
-    if (x < x1)
-        x = (-xIntermediate - 1.0f) / c;
-    return copysign (x, y);
-}
-
-
-
-template<class Rtype>
-static bool
-rangecompress_ (ImageBuf &R, bool useluma, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(rangecompress_<Rtype>, boost::ref(R), useluma,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    const ImageSpec &Rspec (R.spec());
-    int alpha_channel = Rspec.alpha_channel;
-    int z_channel = Rspec.z_channel;
-    if (roi.nchannels() < 3 ||
-        (alpha_channel >= roi.chbegin && alpha_channel < roi.chbegin+3) ||
-        (z_channel >= roi.chbegin && z_channel < roi.chbegin+3)) {
-        useluma = false;  // No way to use luma
-    }
-
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    for (ImageBuf::Iterator<Rtype> r (R, roi);  !r.done();  ++r) {
-        if (useluma) {
-            float luma = 0.21264f * r[roi.chbegin] + 0.71517f * r[roi.chbegin+1] + 0.07219f * r[roi.chbegin+2];
-            if (fabsf(luma) <= 1.0f)
-                continue;  // Not HDR, no range compression needed
-            float scale = rangecompress (luma) / luma;
-            for (int c = roi.chbegin; c < roi.chend; ++c) {
-                if (c == alpha_channel || c == z_channel)
-                    continue;
-                r[c] = r[c] * scale;
-            }
-        } else {
-            for (int c = roi.chbegin; c < roi.chend; ++c) {
-                if (c == alpha_channel || c == z_channel)
-                    continue;
-                r[c] = rangecompress (r[c]);
-            }
-        }
-    }
-    return true;
-}
-
-
-
-template<class Rtype>
-static bool
-rangeexpand_ (ImageBuf &R, bool useluma, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(rangeexpand_<Rtype>, boost::ref(R), useluma,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    const ImageSpec &Rspec (R.spec());
-    int alpha_channel = Rspec.alpha_channel;
-    int z_channel = Rspec.z_channel;
-    if (roi.nchannels() < 3 ||
-        (alpha_channel >= roi.chbegin && alpha_channel < roi.chbegin+3) ||
-        (z_channel >= roi.chbegin && z_channel < roi.chbegin+3)) {
-        useluma = false;  // No way to use luma
-    }
-
-    ImageBuf::Iterator<Rtype> r (R, roi);
-    for (ImageBuf::Iterator<Rtype> r (R, roi);  !r.done();  ++r) {
-        if (useluma) {
-            float luma = 0.21264f * r[roi.chbegin] + 0.71517f * r[roi.chbegin+1] + 0.07219f * r[roi.chbegin+2];
-            if (fabsf(luma) <= 1.0f)
-                continue;  // Not HDR, no range compression needed
-            float scale = rangeexpand (luma) / luma;
-            for (int c = roi.chbegin; c < roi.chend; ++c) {
-                if (c == alpha_channel || c == z_channel)
-                    continue;
-                r[c] = r[c] * scale;
-            }
-        } else {
-            for (int c = roi.chbegin; c < roi.chend; ++c) {
-                if (c == alpha_channel || c == z_channel)
-                    continue;
-                r[c] = rangeexpand (r[c]);
-            }
-        }
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::rangecompress (ImageBuf &dst, bool useluma,
-                             ROI roi, int nthreads)
-{
-    // If the data type can't handle extended range, this is a no-op
-    int basetype = dst.spec().format.basetype;
-    if (basetype != TypeDesc::FLOAT && basetype != TypeDesc::HALF &&
-        basetype != TypeDesc::DOUBLE)
-        return true;
-
-    IBAprep (roi, &dst);
-    switch (basetype) {
-    case TypeDesc::FLOAT:
-        return rangecompress_<float> (dst, useluma, roi, nthreads);
-    case TypeDesc::HALF:
-        return rangecompress_<half> (dst, useluma, roi, nthreads);
-    case TypeDesc::DOUBLE:
-        return rangecompress_<double> (dst, useluma, roi, nthreads);
-    default:
-        return true;
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::rangeexpand (ImageBuf &dst, bool useluma,
-                           ROI roi, int nthreads)
-{
-    // If the data type can't handle extended range, this is a no-op
-    int basetype = dst.spec().format.basetype;
-    if (basetype != TypeDesc::FLOAT && basetype != TypeDesc::HALF &&
-        basetype != TypeDesc::DOUBLE)
-        return true;
-
-    IBAprep (roi, &dst);
-    switch (basetype) {
-    case TypeDesc::FLOAT:
-        return rangeexpand_<float> (dst, useluma, roi, nthreads);
-    case TypeDesc::HALF:
-        return rangeexpand_<half> (dst, useluma, roi, nthreads);
-    case TypeDesc::DOUBLE:
-        return rangeexpand_<double> (dst, useluma, roi, nthreads);
-    default:
-        return true;
-    }
-    return true;
-}
-
-
-
-template<class Rtype>
-static bool
-unpremult_ (ImageBuf &R, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(unpremult_<Rtype>, boost::ref(R),
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    int alpha_channel = R.spec().alpha_channel;
-    int z_channel = R.spec().z_channel;
-    for (ImageBuf::Iterator<Rtype> r (R, roi);  !r.done();  ++r) {
-        float alpha = r[alpha_channel];
-        if (alpha == 0.0f || alpha == 1.0f)
-            continue;
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            if (c != alpha_channel && c != z_channel)
-                r[c] = r[c] / alpha;
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::unpremult (ImageBuf &dst,  ROI roi, int nthreads)
-{
-    if (dst.spec().alpha_channel < 0)
-        return true;
-
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES ("unpremult", unpremult_, dst.spec().format,
-                         dst, roi, nthreads);
-    return true;
-}
-
-
-
-template<class Rtype>
-static bool
-premult_ (ImageBuf &R, ROI roi, int nthreads)
-{
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Possible multiple thread case -- recurse via parallel_image
-        ImageBufAlgo::parallel_image (
-            boost::bind(premult_<Rtype>, boost::ref(R),
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    int alpha_channel = R.spec().alpha_channel;
-    int z_channel = R.spec().z_channel;
-    for (ImageBuf::Iterator<Rtype> r (R, roi);  !r.done();  ++r) {
-        float alpha = r[alpha_channel];
-        if (alpha == 1.0f)
-            continue;
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            if (c != alpha_channel && c != z_channel)
-                r[c] = r[c] * alpha;
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::premult (ImageBuf &dst, ROI roi, int nthreads)
-{
-    if (dst.spec().alpha_channel < 0)
-        return true;
-
-    IBAprep (roi, &dst);
-    OIIO_DISPATCH_TYPES ("premult", premult_, dst.spec().format,
-                         dst, roi, nthreads);
-    return true;
-}
-
-
-
-inline void
-reset (ImageBufAlgo::PixelStats &p, int nchannels)
-{
-    const float inf = std::numeric_limits<float>::infinity();
-    p.min.clear ();          p.min.resize (nchannels, inf);
-    p.max.clear ();          p.max.resize (nchannels, -inf);
-    p.avg.clear ();          p.avg.resize (nchannels);
-    p.stddev.clear ();       p.stddev.resize (nchannels);
-    p.nancount.clear ();     p.nancount.resize (nchannels, 0);
-    p.infcount.clear ();     p.infcount.resize (nchannels, 0);
-    p.finitecount.clear ();  p.finitecount.resize (nchannels, 0);
-    p.sum.clear ();          p.sum.resize (nchannels, 0.0);
-    p.sum2.clear ();         p.sum2.resize (nchannels, 0.0);
-}
-
-
-inline void
-merge (ImageBufAlgo::PixelStats &sum, const ImageBufAlgo::PixelStats &p)
-{
-    ASSERT (sum.min.size() == p.min.size());
-    for (size_t c = 0, e = sum.min.size(); c < e;  ++c) {
-        sum.min[c] = std::min (sum.min[c], p.min[c]);
-        sum.max[c] = std::max (sum.max[c], p.max[c]);
-        sum.nancount[c] += p.nancount[c];
-        sum.infcount[c] += p.infcount[c];
-        sum.finitecount[c] += p.finitecount[c];
-        sum.sum[c] += p.sum[c];
-        sum.sum2[c] += p.sum2[c];
-    }
-}
-
-
-inline void
-val (ImageBufAlgo::PixelStats &p, int c, float value)
-{
-    if (isnan (value)) {
-        ++p.nancount[c];
-        return;
-    }
-    if (isinf (value)) {
-        ++p.infcount[c];
-        return;
-    }
-    ++p.finitecount[c];
-    p.sum[c] += value;
-    p.sum2[c] += value*value;
-    p.min[c] = std::min (value, p.min[c]);
-    p.max[c] = std::max (value, p.max[c]);
-}
-
-
-
-inline void
-finalize (ImageBufAlgo::PixelStats &p)
-{
-    for (size_t c = 0, e = p.min.size();  c < e;  ++c) {
-        if (p.finitecount[c] == 0) {
-            p.min[c] = 0.0;
-            p.max[c] = 0.0;
-            p.avg[c] = 0.0;
-            p.stddev[c] = 0.0;
-        } else {
-            double Count = static_cast<double>(p.finitecount[c]);
-            double davg = p.sum[c] / Count;
-            p.avg[c] = static_cast<float>(davg);
-            p.stddev[c] = static_cast<float>(sqrt(p.sum2[c]/Count - davg*davg));
-        }
-    }
-}
-
-
-
-template <class T>
-static bool
-computePixelStats_ (const ImageBuf &src, ImageBufAlgo::PixelStats &stats,
-                    ROI roi, int nthreads)
-{
-    if (! roi.defined())
-        roi = get_roi (src.spec());
-    else
-        roi.chend = std::min (roi.chend, src.nchannels());
-
-    int nchannels = src.spec().nchannels;
-
-    // Use local storage for smaller batches, then merge the batches
-    // into the final results.  This preserves precision for large
-    // images, where the running total may be too big to incorporate the
-    // contributions of individual pixel values without losing
-    // precision.
-    //
-    // This approach works best when the batch size is the sqrt of
-    // numpixels, which makes the num batches roughly equal to the
-    // number of pixels / batch.
-    ImageBufAlgo::PixelStats tmp;
-    reset (tmp, nchannels);
-    reset (stats, nchannels);
-    
-    int PIXELS_PER_BATCH = std::max (1024,
-            static_cast<int>(sqrt((double)src.spec().image_pixels())));
-    
-    if (src.deep()) {
-        // Loop over all pixels ...
-        for (ImageBuf::ConstIterator<T> s(src, roi); ! s.done();  ++s) {
-            int samples = s.deep_samples();
-            if (! samples)
-                continue;
-            for (int c = roi.chbegin;  c < roi.chend;  ++c) {
-                for (int i = 0;  i < samples;  ++i) {
-                    float value = s.deep_value (c, i);
-                    val (tmp, c, value);
-                    if ((tmp.finitecount[c] % PIXELS_PER_BATCH) == 0) {
-                        merge (stats, tmp);
-                        reset (tmp, nchannels);
-                    }
-                }
-            }
-        }
-    } else {  // Non-deep case
-        // Loop over all pixels ...
-        for (ImageBuf::ConstIterator<T> s(src, roi); ! s.done();  ++s) {
-            for (int c = roi.chbegin;  c < roi.chend;  ++c) {
-                float value = s[c];
-                val (tmp, c, value);
-                if ((tmp.finitecount[c] % PIXELS_PER_BATCH) == 0) {
-                    merge (stats, tmp);
-                    reset (tmp, nchannels);
-                }
-            }
-        }
-    }
-
-    // Merge anything left over
-    merge (stats, tmp);
-
-    // Compute final results
-    finalize (stats);
-    
-    return true;
-};
-
-
-
-bool
-ImageBufAlgo::computePixelStats (PixelStats &stats, const ImageBuf &src,
-                                 ROI roi, int nthreads)
-{
-    if (! roi.defined())
-        roi = get_roi (src.spec());
-    else
-        roi.chend = std::min (roi.chend, src.nchannels());
-    int nchannels = src.spec().nchannels;
-    if (nchannels == 0) {
-        src.error ("%d-channel images not supported", nchannels);
-        return false;
-    }
-
-    OIIO_DISPATCH_TYPES ("computePixelStats", computePixelStats_,
-                         src.spec().format, src, stats, roi, nthreads);
-    return false;
-}
-
-
-
-template<class BUFT>
-inline void
-compare_value (ImageBuf::ConstIterator<BUFT,float> &a, int chan,
-               float aval, float bval, ImageBufAlgo::CompareResults &result,
-               float &maxval, double &batcherror, double &batch_sqrerror,
-               bool &failed, bool &warned, float failthresh, float warnthresh)
-{
-    maxval = std::max (maxval, std::max (aval, bval));
-    double f = fabs (aval - bval);
-    batcherror += f;
-    batch_sqrerror += f*f;
-    if (f > result.maxerror) {
-        result.maxerror = f;
-        result.maxx = a.x();
-        result.maxy = a.y();
-        result.maxz = a.z();
-        result.maxc = chan;
-    }
-    if (! warned && f > warnthresh) {
-        ++result.nwarn;
-        warned = true;
-    }
-    if (! failed && f > failthresh) {
-        ++result.nfail;
-        failed = true;
-    }
-}
-
-
-
-template <class Atype, class Btype>
-static bool
-compare_ (const ImageBuf &A, const ImageBuf &B,
-          float failthresh, float warnthresh,
-          ImageBufAlgo::CompareResults &result,
-          ROI roi, int nthreads)
-{
-    imagesize_t npels = roi.npixels();
-    imagesize_t nvals = npels * roi.nchannels();
-    int Achannels = A.nchannels(), Bchannels = B.nchannels();
-
-    // Compare the two images.
-    //
-    double totalerror = 0;
-    double totalsqrerror = 0;
-    result.maxerror = 0;
-    result.maxx=0, result.maxy=0, result.maxz=0, result.maxc=0;
-    result.nfail = 0, result.nwarn = 0;
-    float maxval = 1.0;  // max possible value
-
-    ImageBuf::ConstIterator<Atype> a (A, roi, ImageBuf::WrapBlack);
-    ImageBuf::ConstIterator<Btype> b (B, roi, ImageBuf::WrapBlack);
-    bool deep = A.deep();
-    // Break up into batches to reduce cancelation errors as the error
-    // sums become too much larger than the error for individual pixels.
-    const int batchsize = 4096;   // As good a guess as any
-    for ( ;  ! a.done();  ) {
-        double batcherror = 0;
-        double batch_sqrerror = 0;
-        if (deep) {
-            for (int i = 0;  i < batchsize && !a.done();  ++i, ++a, ++b) {
-                bool warned = false, failed = false;  // For this pixel
-                for (int c = roi.chbegin;  c < roi.chend;  ++c)
-                    for (int s = 0, e = a.deep_samples(); s < e;  ++s) {
-                        compare_value (a, c, a.deep_value(c,s),
-                                       b.deep_value(c,s), result, maxval,
-                                       batcherror, batch_sqrerror,
-                                       failed, warned, failthresh, warnthresh);
-                    }
-            }
-        } else {  // non-deep
-            for (int i = 0;  i < batchsize && !a.done();  ++i, ++a, ++b) {
-                bool warned = false, failed = false;  // For this pixel
-                for (int c = roi.chbegin;  c < roi.chend;  ++c)
-                    compare_value (a, c, c < Achannels ? a[c] : 0.0f,
-                                   c < Bchannels ? b[c] : 0.0f,
-                                   result, maxval, batcherror, batch_sqrerror,
-                                   failed, warned, failthresh, warnthresh);
-            }
-        }
-        totalerror += batcherror;
-        totalsqrerror += batch_sqrerror;
-    }
-    result.meanerror = totalerror / nvals;
-    result.rms_error = sqrt (totalsqrerror / nvals);
-    result.PSNR = 20.0 * log10 (maxval / result.rms_error);
-    return result.nfail == 0;
-}
-
-
-
-bool
-ImageBufAlgo::compare (const ImageBuf &A, const ImageBuf &B,
-                       float failthresh, float warnthresh,
-                       ImageBufAlgo::CompareResults &result,
-                       ROI roi, int nthreads)
-{
-    // If no ROI is defined, use the union of the data windows of the two
-    // images.
-    if (! roi.defined())
-        roi = roi_union (get_roi(A.spec()), get_roi(B.spec()));
-    roi.chend = std::min (roi.chend, std::max(A.nchannels(), B.nchannels()));
-
-    // Deep and non-deep images cannot be compared
-    if (B.deep() != A.deep())
-        return false;
-
-    OIIO_DISPATCH_TYPES2 ("compare", compare_,
-                          A.spec().format, B.spec().format,
-                          A, B, failthresh, warnthresh, result,
-                          roi, nthreads);
-    // FIXME - The nthreads argument is for symmetry with the rest of
-    // ImageBufAlgo and for future expansion. But for right now, we
-    // don't actually split by threads.  Maybe later.
-    return false;
-}
-
-
-
-template<typename T>
-static inline bool
-isConstantColor_ (const ImageBuf &src, float *color,
-                  ROI roi, int nthreads)
-{
-    // Iterate using the native typing (for speed).
-    std::vector<T> constval (roi.nchannels());
-    ImageBuf::ConstIterator<T,T> s (src, roi);
-    for (int c = roi.chbegin;  c < roi.chend;  ++c)
-        constval[c] = s[c];
-
-    // Loop over all pixels ...
-    for ( ; ! s.done();  ++s) {
-        for (int c = roi.chbegin;  c < roi.chend;  ++c)
-            if (constval[c] != s[c])
-                return false;
-    }
-    
-    if (color) {
-        ImageBuf::ConstIterator<T,float> s (src, roi);
-        for (int c = 0;  c < roi.chbegin; ++c)
-            color[c] = 0.0f;
-        for (int c = roi.chbegin; c < roi.chend; ++c)
-            color[c] = s[c];
-        for (int c = roi.chend;  c < src.nchannels(); ++c)
-            color[c] = 0.0f;
-    }
-
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::isConstantColor (const ImageBuf &src, float *color,
-                               ROI roi, int nthreads)
-{
-    // If no ROI is defined, use the data window of src.
-    if (! roi.defined())
-        roi = get_roi(src.spec());
-    roi.chend = std::min (roi.chend, src.nchannels());
-
-    if (roi.nchannels() == 0)
-        return true;
-    
-    OIIO_DISPATCH_TYPES ("isConstantColor", isConstantColor_,
-                         src.spec().format, src, color, roi, nthreads);
-    // FIXME -  The nthreads argument is for symmetry with the rest of
-    // ImageBufAlgo and for future expansion. But for right now, we
-    // don't actually split by threads.  Maybe later.
-};
-
-
-
-template<typename T>
-static inline bool
-isConstantChannel_ (const ImageBuf &src, int channel, float val,
-                    ROI roi, int nthreads)
-{
-
-    T v = convert_type<float,T> (val);
-    for (ImageBuf::ConstIterator<T,T> s(src, roi);  !s.done();  ++s)
-        if (s[channel] != v)
-            return false;
-    return true;
-}
-
-
-bool
-ImageBufAlgo::isConstantChannel (const ImageBuf &src, int channel, float val,
-                                 ROI roi, int nthreads)
-{
-    // If no ROI is defined, use the data window of src.
-    if (! roi.defined())
-        roi = get_roi(src.spec());
-
-    if (channel < 0 || channel >= src.nchannels())
-        return false;  // that channel doesn't exist in the image
-
-    OIIO_DISPATCH_TYPES ("isConstantChannel", isConstantChannel_,
-                         src.spec().format, src, channel, val, roi, nthreads);
-    // FIXME -  The nthreads argument is for symmetry with the rest of
-    // ImageBufAlgo and for future expansion. But for right now, we
-    // don't actually split by threads.  Maybe later.
-};
-
-
-
-template<typename T>
-static inline bool
-isMonochrome_ (const ImageBuf &src, ROI roi, int nthreads)
-{
-    int nchannels = src.nchannels();
-    if (nchannels < 2) return true;
-    
-    // Loop over all pixels ...
-    for (ImageBuf::ConstIterator<T,T> s(src, roi);  ! s.done();  ++s) {
-        T constvalue = s[roi.chbegin];
-        for (int c = roi.chbegin+1;  c < roi.chend;  ++c)
-            if (s[c] != constvalue)
-                return false;
-    }
-    return true;
-}
-
-
-
-bool
-ImageBufAlgo::isMonochrome (const ImageBuf &src, ROI roi, int nthreads)
-{
-    // If no ROI is defined, use the data window of src.
-    if (! roi.defined())
-        roi = get_roi(src.spec());
-    roi.chend = std::min (roi.chend, src.nchannels());
-    if (roi.nchannels() < 2)
-        return true;  // 1 or fewer channels are always "monochrome"
-
-    OIIO_DISPATCH_TYPES ("isMonochrome", isMonochrome_, src.spec().format,
-                         src, roi, nthreads);
-    // FIXME -  The nthreads argument is for symmetry with the rest of
-    // ImageBufAlgo and for future expansion. But for right now, we
-    // don't actually split by threads.  Maybe later.
-};
-
-
-
-namespace {
-
-std::string
-simplePixelHashSHA1 (const ImageBuf &src,
-                     const std::string & extrainfo, ROI roi)
-{
-    if (! roi.defined())
-        roi = get_roi (src.spec());
-
-    bool localpixels = src.localpixels();
-    imagesize_t scanline_bytes = roi.width() * src.spec().pixel_bytes();
-    ASSERT (scanline_bytes < std::numeric_limits<unsigned int>::max());
-    // Do it a few scanlines at a time
-    int chunk = std::max (1, int(16*1024*1024/scanline_bytes));
-
-    std::vector<unsigned char> tmp;
-    if (! localpixels)
-        tmp.resize (chunk*scanline_bytes);
-
-#ifdef USE_OPENSSL
-    // If OpenSSL was available at build time, use its SHA-1
-    // implementation, which is about 20% faster than CSHA1.
-    SHA_CTX sha;
-    SHA1_Init (&sha);
-
-    for (int z = roi.zbegin, zend=roi.zend;  z < zend;  ++z) {
-        for (int y = roi.ybegin, yend=roi.yend;  y < yend;  y += chunk) {
-            int y1 = std::min (y+chunk, yend);
-            if (localpixels) {
-                SHA1_Update (&sha, src.pixeladdr (roi.xbegin, y, z),
-                            (unsigned int) scanline_bytes*(y1-y));
-            } else {
-                src.get_pixels (roi.xbegin, roi.xend, y, y1, z, z+1,
-                                src.spec().format, &tmp[0]);
-                SHA1_Update (&sha, &tmp[0], (unsigned int) scanline_bytes*(y1-y));
-            }
-        }
-    }
-    
-    // If extra info is specified, also include it in the sha computation
-    if (!extrainfo.empty())
-        SHA1_Update (&sha, extrainfo.c_str(), extrainfo.size());
-
-    unsigned char md[SHA_DIGEST_LENGTH];
-    char hash_digest[2*SHA_DIGEST_LENGTH+1];
-    SHA1_Final (md, &sha);
-    for (int i = 0;  i < SHA_DIGEST_LENGTH;  ++i)
-        sprintf (hash_digest+2*i, "%02X", (int)md[i]);
-    hash_digest[2*SHA_DIGEST_LENGTH] = 0;
-    return std::string (hash_digest);
-    
-#else
-    // Fall back on CSHA1 if OpenSSL was not available or if 
-    CSHA1 sha;
-    sha.Reset ();
-    
-    for (int z = roi.zbegin, zend=roi.zend;  z < zend;  ++z) {
-        for (int y = roi.ybegin, yend=roi.yend;  y < yend;  y += chunk) {
-            int y1 = std::min (y+chunk, yend);
-            if (localpixels) {
-                sha.Update ((const unsigned char *)src.pixeladdr (roi.xbegin, y, z),
-                            (unsigned int) scanline_bytes*(y1-y));
-            } else {
-                src.get_pixels (roi.xbegin, roi.xend, y, y1, z, z+1,
-                                src.spec().format, &tmp[0]);
-                sha.Update (&tmp[0], (unsigned int) scanline_bytes*(y1-y));
-            }
-        }
-    }
-    
-    // If extra info is specified, also include it in the sha computation
-    if (!extrainfo.empty()) {
-        sha.Update ((const unsigned char*) extrainfo.c_str(), extrainfo.size());
-    }
-    
-    sha.Final ();
-    std::string hash_digest;
-    sha.ReportHashStl (hash_digest, CSHA1::REPORT_HEX_SHORT);
-
-    return hash_digest;
-#endif
-}
-
-
-
-// Wrapper to single-threadedly SHA1 hash a region in blocks and store
-// the results in a designated place.
-static void
-sha1_hasher (const ImageBuf *src, ROI roi, int blocksize,
-             std::string *results, int firstresult)
-{
-    ROI broi = roi;
-    for (int y = roi.ybegin; y < roi.yend; y += blocksize) {
-        broi.ybegin = y;
-        broi.yend = std::min (y+blocksize, roi.yend);
-        std::string s = simplePixelHashSHA1 (*src, "", broi);
-        results[firstresult++] = s;
-    }
-}
-
-} // anon namespace
-
-
-
-std::string
-ImageBufAlgo::computePixelHashSHA1 (const ImageBuf &src,
-                                    const std::string & extrainfo,
-                                    ROI roi, int blocksize, int nthreads)
-{
-    if (! roi.defined())
-        roi = get_roi (src.spec());
-
-    // Fall back to whole-image hash for only one block
-    if (blocksize <= 0 || blocksize >= roi.height())
-        return simplePixelHashSHA1 (src, extrainfo, roi);
-
-    // Request for 0 threads means "use the OIIO global thread count"
-    if (nthreads <= 0)
-        OIIO::getattribute ("threads", nthreads);
-
-    int nblocks = (roi.height()+blocksize-1) / blocksize;
-    std::vector<std::string> results (nblocks);
-    if (nthreads <= 1) {
-        sha1_hasher (&src, roi, blocksize, &results[0], 0);
-    } else {
-        // parallel case
-        boost::thread_group threads;
-        int blocks_per_thread = (nblocks+nthreads-1) / nthreads;
-        ROI broi = roi;
-        for (int b = 0, t = 0;  b < nblocks;  b += blocks_per_thread, ++t) {
-            int y = roi.ybegin + b*blocksize;
-            if (y >= roi.yend)
-                break;
-            broi.ybegin = y;
-            broi.yend = std::min (y+blocksize*blocks_per_thread, roi.yend);
-            threads.add_thread (new boost::thread (sha1_hasher, &src, broi,
-                                                   blocksize, &results[0], b));
-        }
-        threads.join_all ();
-    }
-
-#ifdef USE_OPENSSL
-    // If OpenSSL was available at build time, use its SHA-1
-    // implementation, which is about 20% faster than CSHA1.
-    SHA_CTX sha;
-    SHA1_Init (&sha);
-    for (int b = 0;  b < nblocks;  ++b)
-        SHA1_Update (&sha, results[b].c_str(), results[b].size());
-    if (extrainfo.size())
-        SHA1_Update (&sha, extrainfo.c_str(), extrainfo.size());
-    unsigned char md[SHA_DIGEST_LENGTH];
-    char hash_digest[2*SHA_DIGEST_LENGTH+1];
-    SHA1_Final (md, &sha);
-    for (int i = 0;  i < SHA_DIGEST_LENGTH;  ++i)
-        sprintf (hash_digest+2*i, "%02X", (int)md[i]);
-    hash_digest[2*SHA_DIGEST_LENGTH] = 0;
-    return std::string (hash_digest);
-#else
-    // Fall back on CSHA1 if OpenSSL was not available or if 
-    CSHA1 sha;
-    sha.Reset ();
-    for (int b = 0;  b < nblocks;  ++b)
-        sha.Update ((const unsigned char *)results[b].c_str(), results[b].size());
-    if (extrainfo.size())
-        sha.Update ((const unsigned char *)extrainfo.c_str(), extrainfo.size());
-    sha.Final ();
-    std::string hash_digest;
-    sha.ReportHashStl (hash_digest, CSHA1::REPORT_HEX_SHORT);
-    return hash_digest;
-#endif
-}
-
 
 
 
@@ -2277,438 +737,6 @@ ImageBufAlgo::unsharp_mask (ImageBuf &dst, const ImageBuf &src,
 
 
 
-
-namespace
-{
-
-// Make sure isfinite is defined for 'half'
-inline bool isfinite (half h) { return h.isFinite(); }
-
-
-template<typename T>
-bool fixNonFinite_ (ImageBuf &dst, ImageBufAlgo::NonFiniteFixMode mode,
-                    int *pixelsFixed, ROI roi, int nthreads)
-{
-    if (mode != ImageBufAlgo::NONFINITE_NONE &&
-        mode != ImageBufAlgo::NONFINITE_BLACK &&
-        mode != ImageBufAlgo::NONFINITE_BOX3) {
-        // Something went wrong
-        dst.error ("fixNonFinite: unknown repair mode");
-        return false;
-    }
-
-    if (nthreads != 1 && roi.npixels() >= 1000) {
-        // Lots of pixels and request for multi threads? Parallelize.
-        ImageBufAlgo::parallel_image (
-            boost::bind(fixNonFinite_<T>, boost::ref(dst), mode, pixelsFixed,
-                        _1 /*roi*/, 1 /*nthreads*/),
-            roi, nthreads);
-        return true;
-    }
-
-    // Serial case
-
-    ROI dstroi = get_roi (dst.spec());
-    int count = 0;   // Number of pixels with nonfinite values
-
-    if (mode == ImageBufAlgo::NONFINITE_NONE) {
-        // Just count the number of pixels with non-finite values
-        for (ImageBuf::Iterator<T,T> pixel (dst);  ! pixel.done();  ++pixel) {
-            for (int c = roi.chbegin;  c < roi.chend;  ++c) {
-                T value = pixel[c];
-                if (! isfinite(value)) {
-                    ++count;
-                    break;  // only count one per pixel
-                }
-            }
-        }
-    } else if (mode == ImageBufAlgo::NONFINITE_BLACK) {
-        // Replace non-finite pixels with black
-        for (ImageBuf::Iterator<T,T> pixel (dst);  ! pixel.done();  ++pixel) {
-            bool fixed = false;
-            for (int c = roi.chbegin;  c < roi.chend;  ++c) {
-                T value = pixel[c];
-                if (! isfinite(value)) {
-                    pixel[c] = T(0.0);
-                    fixed = true;
-                }
-            }
-            if (fixed)
-                ++count;
-        }
-    } else if (mode == ImageBufAlgo::NONFINITE_BOX3) {
-        // Replace non-finite pixels with a simple 3x3 window average
-        // (the average excluding non-finite pixels, of course)
-        for (ImageBuf::Iterator<T,T> pixel (dst);  ! pixel.done();  ++pixel) {
-            bool fixed = false;
-            for (int c = roi.chbegin;  c < roi.chend;  ++c) {
-                T value = pixel[c];
-                if (! isfinite (value)) {
-                    int numvals = 0;
-                    T sum (0.0);
-                    ROI roi2 (pixel.x()-1, pixel.x()+2,
-                              pixel.y()-1, pixel.y()+2,
-                              pixel.z()-1, pixel.z()+2);
-                    roi2 = roi_intersection (roi2, dstroi);
-                    for (ImageBuf::Iterator<T,T> i(dst,roi2); !i.done(); ++i) {
-                        T v = i[c];
-                        if (isfinite (v)) {
-                            sum += v;
-                            ++numvals;
-                        }
-                    }
-                    pixel[c] = numvals ? T(sum / numvals) : T(0.0);
-                    fixed = true;
-                }
-            }
-            if (fixed)
-                ++count;
-        }
-    }
-    
-    if (pixelsFixed) {
-        // Update pixelsFixed atomically -- that's what makes this whole
-        // function thread-safe.
-        *(atomic_int *)pixelsFixed += count;
-    }
-
-    return true;
-}
-
-} // anon namespace
-
-
-
-/// Fix all non-finite pixels (nan/inf) using the specified approach
-bool
-ImageBufAlgo::fixNonFinite (ImageBuf &src, 
-                            NonFiniteFixMode mode, int *pixelsFixed,
-                            ROI roi, int nthreads)
-{
-    // If no ROI is defined, use the data window of src.
-    if (! roi.defined())
-        roi = get_roi(src.spec());
-    roi.chend = std::min (roi.chend, src.nchannels());
-
-    // Initialize
-    if (pixelsFixed)
-        *pixelsFixed = 0;
-
-    switch (src.spec().format.basetype) {
-    case TypeDesc::FLOAT :
-        return fixNonFinite_<float> (src, mode, pixelsFixed, roi, nthreads);
-    case TypeDesc::HALF  :
-        return fixNonFinite_<half> (src, mode, pixelsFixed, roi, nthreads);
-    case TypeDesc::DOUBLE:
-        return fixNonFinite_<double> (src, mode, pixelsFixed, roi, nthreads);
-    default:
-        // All other format types aren't capable of having nonfinite
-        // pixel values.
-        return true;
-    }
-}
-
-
-
-// DEPRECATED 2-argument version
-bool
-ImageBufAlgo::fixNonFinite (ImageBuf &dst, const ImageBuf &src,
-                            NonFiniteFixMode mode, int *pixelsFixed)
-{
-    ROI roi;
-    IBAprep (roi, &dst, &src);
-    if (dst.nchannels() != src.nchannels()) {
-        dst.error ("channel number mismatch: %d vs. %d", 
-                   dst.spec().nchannels, src.spec().nchannels);
-        return false;
-    }
-    if ((const ImageBuf *)&dst != &src)
-        if (! dst.copy (src))
-            return false;
-    return fixNonFinite (dst, mode, pixelsFixed, roi);
-}
-
-
-namespace {   // anonymous namespace
-
-static bool
-decode_over_channels (const ImageBuf &R, int &nchannels, 
-                      int &alpha, int &z, int &colors)
-{
-    if (! R.initialized()) {
-        alpha = -1;
-        z = -1;
-        colors = 0;
-        return false;
-    }
-    const ImageSpec &spec (R.spec());
-    alpha =  spec.alpha_channel;
-    bool has_alpha = (alpha >= 0);
-    z = spec.z_channel;
-    bool has_z = (z >= 0);
-    nchannels = spec.nchannels;
-    colors = nchannels - has_alpha - has_z;
-    if (! has_alpha && colors == 4) {
-        // No marked alpha channel, but suspiciously 4 channel -- assume
-        // it's RGBA. 
-        has_alpha = true;
-        colors -= 1;
-        // Assume alpha is the highest channel that's not z
-        alpha = nchannels - 1;
-        if (alpha == z)
-            --alpha;
-    }
-    return true;
-}
-
-
-
-// Fully type-specialized version of over.
-template<class Rtype, class Atype, class Btype>
-static bool
-over_impl (ImageBuf &R, const ImageBuf &A, const ImageBuf &B, ROI roi,
-           bool zcomp=false, bool z_zeroisinf=false)
-{
-    if (R.spec().format != BaseTypeFromC<Rtype>::value ||
-        A.spec().format != BaseTypeFromC<Atype>::value ||
-        B.spec().format != BaseTypeFromC<Btype>::value) {
-        R.error ("Unsupported pixel data format combination '%s / %s / %s'",
-                 R.spec().format, A.spec().format, B.spec().format);
-        return false;   // double check that types match
-    }
-
-    // It's already guaranteed that R, A, and B have matching channel
-    // ordering, and have an alpha channel.  So just decode one.
-    int nchannels = 0, alpha_channel = 0, z_channel = 0, ncolor_channels = 0;
-    decode_over_channels (R, nchannels, alpha_channel,
-                          z_channel, ncolor_channels);
-    bool has_z = (z_channel >= 0);
-
-    ImageBuf::ConstIterator<Atype, float> a (A);
-    ImageBuf::ConstIterator<Btype, float> b (B);
-    ImageBuf::Iterator<Rtype, float> r (R, roi);
-    for ( ; ! r.done(); r++) {
-        a.pos (r.x(), r.y(), r.z());
-        b.pos (r.x(), r.y(), r.z());
-
-        if (! a.exists()) {
-            if (! b.exists()) {
-                // a and b outside their data window -- "empty" pixels
-                for (int c = 0; c < nchannels; c++)
-                    r[c] = 0.0f;
-            } else {
-                // a doesn't exist, but b does -- copy B
-                for (int c = 0; c < nchannels; ++c)
-                    r[c] = b[c];
-            }
-            continue;
-        }
-
-        if (! b.exists()) {
-            // a exists, b does not -- copy A
-            for (int c = 0; c < nchannels; ++c)
-                r[c] = a[c];
-            continue;
-        }
-
-        // At this point, a and b exist.
-        float az = 0.0f, bz = 0.0f;
-        bool a_is_closer = true;  // will remain true if !zcomp
-        if (zcomp && has_z) {
-            az = a[z_channel];
-            bz = b[z_channel];
-            if (z_zeroisinf) {
-                if (az == 0.0f) az = std::numeric_limits<float>::max();
-                if (bz == 0.0f) bz = std::numeric_limits<float>::max();
-            }
-            a_is_closer = (az <= bz);
-        }
-        if (a_is_closer) {
-            // A over B
-            float alpha = clamp (a[alpha_channel], 0.0f, 1.0f);
-            float one_minus_alpha = 1.0f - alpha;
-            for (int c = 0;  c < nchannels;  c++)
-                r[c] = a[c] + one_minus_alpha * b[c];
-            if (has_z)
-                r[z_channel] = (alpha != 0.0) ? a[z_channel] : b[z_channel];
-        } else {
-            // B over A -- because we're doing a Z composite
-            float alpha = clamp (b[alpha_channel], 0.0f, 1.0f);
-            float one_minus_alpha = 1.0f - alpha;
-            for (int c = 0;  c < nchannels;  c++)
-                r[c] = b[c] + one_minus_alpha * a[c];
-            r[z_channel] = (alpha != 0.0) ? b[z_channel] : a[z_channel];
-        }
-    }
-    return true;
-}
-
-}    // anonymous namespace
-
-
-bool
-ImageBufAlgo::over (ImageBuf &R, const ImageBuf &A, const ImageBuf &B, ROI roi,
-                    int nthreads)
-{
-    const ImageSpec &specR = R.spec();
-    const ImageSpec &specA = A.spec();
-    const ImageSpec &specB = B.spec();
-
-    int nchannels_R, nchannels_A, nchannels_B;
-    int alpha_R, alpha_A, alpha_B;
-    int z_R, z_A, z_B;
-    int colors_R, colors_A, colors_B;
-    bool initialized_R = decode_over_channels (R, nchannels_R, alpha_R,
-                                               z_R, colors_R);
-    bool initialized_A = decode_over_channels (A, nchannels_A, alpha_A,
-                                               z_A, colors_A);
-    bool initialized_B = decode_over_channels (B, nchannels_B, alpha_B,
-                                               z_B, colors_B);
-
-    if (! initialized_A || ! initialized_B) {
-        R.error ("Can't 'over' uninitialized images");
-        return false;
-    }
-
-    // Fail if the input images don't have an alpha channel.
-    if (alpha_A < 0 || alpha_B < 0 || (initialized_R && alpha_R < 0)) {
-        R.error ("'over' requires alpha channels");
-        return false;
-    }
-    // Fail for mismatched channel counts
-    if (colors_A != colors_B || colors_A < 1) {
-        R.error ("Can't 'over' images with mismatched color channel counts (%d vs %d)",
-                 colors_A, colors_B);
-        return false;
-    }
-    // Fail for unaligned alpha or z channels
-    if (alpha_A != alpha_B || z_A != z_B ||
-        (initialized_R && alpha_R != alpha_A) ||
-        (initialized_R && z_R != z_A)) {
-        R.error ("Can't 'over' images with mismatched channel order",
-                 colors_A, colors_B);
-        return false;
-    }
-    
-    // At present, this operation only supports ImageBuf's containing
-    // float pixel data.
-    if ((initialized_R && specR.format != TypeDesc::TypeFloat) ||
-        specA.format != TypeDesc::TypeFloat ||
-        specB.format != TypeDesc::TypeFloat) {
-        R.error ("Unsupported pixel data format combination '%s = %s over %s'",
-                 specR.format, specA.format, specB.format);
-        return false;
-    }
-
-    // Uninitialized R -> size it to the union of A and B.
-    if (! initialized_R) {
-        ImageSpec newspec = specA;
-        set_roi (newspec, roi_union (get_roi(specA), get_roi(specB)));
-        R.reset ("over", newspec);
-    }
-
-    // Specified ROI -> use it. Unspecified ROI -> initialize from R.
-    if (! roi.defined())
-        roi = get_roi (R.spec());
-
-    parallel_image (boost::bind (over_impl<float,float,float>, boost::ref(R),
-                                 boost::cref(A), boost::cref(B), _1, false, false),
-                    roi, nthreads);
-    return ! R.has_error();
-}
-
-
-
-bool
-ImageBufAlgo::zover (ImageBuf &R, const ImageBuf &A, const ImageBuf &B,
-                     bool z_zeroisinf, ROI roi, int nthreads)
-{
-    const ImageSpec &specR = R.spec();
-    const ImageSpec &specA = A.spec();
-    const ImageSpec &specB = B.spec();
-
-    int nchannels_R, nchannels_A, nchannels_B;
-    int alpha_R, alpha_A, alpha_B;
-    int z_R, z_A, z_B;
-    int colors_R, colors_A, colors_B;
-    bool initialized_R = decode_over_channels (R, nchannels_R, alpha_R,
-                                               z_R, colors_R);
-    bool initialized_A = decode_over_channels (A, nchannels_A, alpha_A,
-                                               z_A, colors_A);
-    bool initialized_B = decode_over_channels (B, nchannels_B, alpha_B,
-                                               z_B, colors_B);
-
-    if (! initialized_A || ! initialized_B) {
-        R.error ("Can't 'zover' uninitialized images");
-        return false;
-    }
-    // Fail if the input images don't have a Z channel.
-    if (z_A < 0 || z_B < 0 || (initialized_R && z_R < 0)) {
-        R.error ("'zover' requires Z channels");
-        return false;
-    }
-    // Fail if the input images don't have an alpha channel.
-    if (alpha_A < 0 || alpha_B < 0 || (initialized_R && alpha_R < 0)) {
-        R.error ("'zover' requires alpha channels");
-        return false;
-    }
-    // Fail for mismatched channel counts
-    if (colors_A != colors_B || colors_A < 1) {
-        R.error ("Can't 'zover' images with mismatched color channel counts (%d vs %d)",
-                 colors_A, colors_B);
-        return false;
-    }
-    // Fail for unaligned alpha or z channels
-    if (alpha_A != alpha_B || z_A != z_B ||
-        (initialized_R && alpha_R != alpha_A) ||
-        (initialized_R && z_R != z_A)) {
-        R.error ("Can't 'zover' images with mismatched channel order",
-                 colors_A, colors_B);
-        return false;
-    }
-    
-    // At present, this operation only supports ImageBuf's containing
-    // float pixel data.
-    if ((initialized_R && specR.format != TypeDesc::TypeFloat) ||
-        specA.format != TypeDesc::TypeFloat ||
-        specB.format != TypeDesc::TypeFloat) {
-        R.error ("Unsupported pixel data format combination '%s = %s zover %s'",
-                 specR.format, specA.format, specB.format);
-        return false;
-    }
-
-    // Uninitialized R -> size it to the union of A and B.
-    if (! initialized_R) {
-        ImageSpec newspec = specA;
-        set_roi (newspec, roi_union (get_roi(specA), get_roi(specB)));
-        R.reset ("zover", newspec);
-    }
-
-    // Specified ROI -> use it. Unspecified ROI -> initialize from R.
-    if (! roi.defined())
-        roi = get_roi (R.spec());
-
-    parallel_image (boost::bind (over_impl<float,float,float>, boost::ref(R),
-                                 boost::cref(A), boost::cref(B), _1,
-                                 true, z_zeroisinf),
-                    roi, nthreads);
-    return ! R.has_error();
-}
-
-
-
-bool
-ImageBufAlgo::zover (ImageBuf &R, const ImageBuf &A, const ImageBuf &B,
-                     ROI roi, int nthreads)
-{
-    // DEPRECATED version -- just call the new version.  This exists to 
-    // avoid breaking link compatibility.  Eventually remove it at the
-    // next major release.
-    return zover (R, A, B, false, roi, nthreads);
-}
-
-
-
 #ifdef USE_FREETYPE
 namespace { // anon
 static mutex ft_mutex;
@@ -2834,150 +862,6 @@ ImageBufAlgo::render_text (ImageBuf &R, int x, int y, const std::string &text,
     R.error ("OpenImageIO was not compiled with FreeType for font rendering");
     return false;   // Font rendering not supported
 #endif
-}
-
-
-namespace { // anonymous namespace
-
-/// histogram_impl -----------------------------------------------------------
-/// Fully type-specialized version of histogram.
-///
-/// Pixel values in min->max range are mapped to 0->(bins-1) range, so that
-/// each value is placed in the appropriate bin. The formula used is:
-/// y = (x-min) * bins/(max-min), where y is the value in the 0->(bins-1)
-/// range and x is the value in the min->max range. There is one special
-/// case x==max for which the formula is not used and x is assigned to the
-/// last bin at position (bins-1) in the vector histogram.
-/// --------------------------------------------------------------------------
-template<class Atype>
-bool
-histogram_impl (const ImageBuf &A, int channel,
-                std::vector<imagesize_t> &histogram, int bins,
-                float min, float max, imagesize_t *submin,
-                imagesize_t *supermax, ROI roi)
-{
-    // Double check A's type.
-    if (A.spec().format != BaseTypeFromC<Atype>::value) {
-        A.error ("Unsupported pixel data format '%s'", A.spec().format);
-        return false;
-    }
-
-    // Initialize.
-    ImageBuf::ConstIterator<Atype, float> a (A, roi);
-    float ratio = bins / (max-min);
-    int bins_minus_1 = bins-1;
-    bool submin_ok = submin != NULL;
-    bool supermax_ok = supermax != NULL;
-    if (submin_ok)
-        *submin = 0;
-    if (supermax_ok)
-        *supermax = 0;
-    histogram.assign(bins, 0);
-
-    // Compute histogram.
-    for ( ; ! a.done(); a++) {
-        float c = a[channel];
-        if (c >= min && c < max) {
-            // Map range min->max to 0->(bins-1).
-            histogram[ (int) ((c-min) * ratio) ]++;
-        } else if (c == max) {
-            histogram[bins_minus_1]++;
-        } else {
-            if (submin_ok && c < min)
-                (*submin)++;
-            else if (supermax_ok)
-                (*supermax)++;
-        }
-    }
-    return true;
-}
-
-} // anonymous namespace
-
-
-
-bool
-ImageBufAlgo::histogram (const ImageBuf &A, int channel,
-                         std::vector<imagesize_t> &histogram, int bins,
-                         float min, float max, imagesize_t *submin,
-                         imagesize_t *supermax, ROI roi)
-{
-    if (A.spec().format != TypeDesc::TypeFloat) {
-        A.error ("Unsupported pixel data format '%s'", A.spec().format);
-        return false;
-    }
-
-    if (A.nchannels() == 0) {
-        A.error ("Input image must have at least 1 channel");
-        return false;
-    }
-
-    if (channel < 0 || channel >= A.nchannels()) {
-        A.error ("Invalid channel %d for input image with channels 0 to %d",
-                  channel, A.nchannels()-1);
-        return false;
-    }
-
-    if (bins < 1) {
-        A.error ("The number of bins must be at least 1");
-        return false;
-    }
-
-    if (max <= min) {
-        A.error ("Invalid range, min must be strictly smaller than max");
-        return false;
-    }
-
-    // Specified ROI -> use it. Unspecified ROI -> initialize from A.
-    if (! roi.defined())
-        roi = get_roi (A.spec());
-
-    histogram_impl<float> (A, channel, histogram, bins, min, max,
-                                  submin, supermax, roi);
-
-    return ! A.has_error();
-}
-
-
-
-bool
-ImageBufAlgo::histogram_draw (ImageBuf &R,
-                              const std::vector<imagesize_t> &histogram)
-{
-    // Fail if there are no bins to draw.
-    int bins = histogram.size();
-    if (bins == 0) {
-        R.error ("There are no bins to draw, the histogram is empty");
-        return false;
-    }
-
-    // Check R and modify it if needed.
-    int height = R.spec().height;
-    if (R.spec().format != TypeDesc::TypeFloat || R.nchannels() != 1 ||
-        R.spec().width != bins) {
-        ImageSpec newspec = ImageSpec (bins, height, 1, TypeDesc::FLOAT);
-        R.reset ("dummy", newspec);
-    }
-
-    // Fill output image R with white color.
-    ImageBuf::Iterator<float, float> r (R);
-    for ( ; ! r.done(); ++r)
-        r[0] = 1;
-
-    // Draw histogram left->right, bottom->up.
-    imagesize_t max = *std::max_element (histogram.begin(), histogram.end());
-    for (int b = 0; b < bins; b++) {
-        int bin_height = (int) ((float)histogram[b]/(float)max*height + 0.5f);
-        if (bin_height != 0) {
-            // Draw one bin at column b.
-            for (int j = 1; j <= bin_height; j++) {
-                int row = height - j;
-                r.pos (b, row);
-                r[0] = 0;
-            }
-        }
-    }
-    return true;
 }
